requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.7.2
attrs==23.1.0
billiard==4.2.0
black==23.12.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
Django==4.0.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.4.2
drf-spectacular==0.27.0
eventlet==0.33.3
frozenlist==1.4.0
greenlet==3.0.2
idna==3.6
inflection==0.5.1
install==1.3.5
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kombu==5.3.4
magic-filter==1.0.12
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
pyshorteners==1.0.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.32.0
requests==2.31.0
rpds-py==0.13.2
six==1.16.0
sqlparse==0.4.4
stripe==7.8.2
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
yarl==1.9.4
pytest~=7.4.3
django-debug-toolbar==4.2.0
